feat: 카테고리 등록/수정/삭제/테스트 API 관리자만 사용 가능하도록 제한

+++ src/main/java/com/capstonebe/capstonebe/category/controller/CategoryController.java
@@ -4,6 +4,10 @@ import com.capstonebe.capstonebe.category.dto.request.CategoryEditRequest;
 import com.capstonebe.capstonebe.category.dto.request.CategoryRegisterRequest;
 import com.capstonebe.capstonebe.category.dto.response.CategoryResponse;
 import com.capstonebe.capstonebe.category.service.CategoryService;
+import com.capstonebe.capstonebe.global.exception.CustomErrorCode;
+import com.capstonebe.capstonebe.global.exception.CustomException;
+import com.capstonebe.capstonebe.security.JwtUtil;
+import com.capstonebe.capstonebe.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -17,23 +21,38 @@ import java.util.List;
 public class CategoryController {
 
     private final CategoryService categoryService;
+    private final UserService userService;
+    private final JwtUtil jwtUtil;
 
     // 카테고리 등록
     @PostMapping("/register")
-    public ResponseEntity<?> registerCategory(@RequestBody @Valid CategoryRegisterRequest request) {
+    public ResponseEntity<?> registerCategory(@RequestBody @Valid CategoryRegisterRequest request,
+                                              @CookieValue(value = "jwt", required = false) String token) {
+
+        validateAdmin(token);
+
         return ResponseEntity.ok(categoryService.registerCategory(request));
     }
 
     // 카테고리 수정
     @PatchMapping("/edit")
-    public ResponseEntity<?> editCategory(@RequestBody @Valid CategoryEditRequest request) {
+    public ResponseEntity<?> editCategory(@RequestBody @Valid CategoryEditRequest request,
+                                          @CookieValue(value = "jwt", required = false) String token) {
+
+        validateAdmin(token);
+
         return ResponseEntity.ok(categoryService.editCategory(request));
     }
 
     // 카테고리 삭제
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteCategory(@PathVariable Long id) {
+    public ResponseEntity<?> deleteCategory(@PathVariable Long id,
+                                            @CookieValue(value = "jwt", required = false) String token) {
+
+        validateAdmin(token);
+
         categoryService.deleteCategory(id);
+
         return ResponseEntity.ok().build();
     }
 
@@ -42,6 +61,13 @@ public class CategoryController {
     public ResponseEntity<?> getAllCategories() {
         List<CategoryResponse> response = categoryService.getAllCategories();
         return ResponseEntity.ok(response);
+    }
+
+    private void validateAdmin(String token) {
+
+        if (token == null || !userService.isAdmin(jwtUtil.extractEmail(token))) {
+            throw new CustomException(CustomErrorCode.IS_NOT_ADMIN);
+        }
     }
